% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_white_shark_trends}
\alias{example_white_shark_trends}
\title{All cached trend series for white shark (\emph{Carcharodon carcharias})}
\usage{
example_white_shark_trends()
}
\value{
A tibble of trend observations from multiple populations.
}
\description{
All cached trend series for white shark (\emph{Carcharodon carcharias})
}
