% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_references}
\alias{parse_references}
\title{Parse bibliographic references listed on a species page}
\usage{
parse_references(doc)
}
\description{
Parse bibliographic references listed on a species page
}
\keyword{internal}
