% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_trends_tables}
\alias{parse_trends_tables}
\title{Parse trend rows embedded in React props}
\usage{
parse_trends_tables(doc)
}
\description{
Parse trend rows embedded in React props
}
\keyword{internal}
