% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkipediaR-package.R
\docType{package}
\name{sharkipediaR-package}
\alias{sharkipediaR}
\alias{sharkipediaR-package}
\title{sharkipediaR: Access Sharkipedia Trait and Trend Data}
\description{
G'day mate! This is a tidyverse-oriented R client for publicly accessible Sharkipedia life-history trait and population trend data. Provides polite HTML retrieval, modular parsing, and reproducible ecological workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecologistpablo/SharkipediaR}
  \item \url{https://ecologistpablo.github.io/SharkipediaR/}
  \item Report bugs at \url{https://github.com/ecologistpablo/SharkipediaR/issues}
}

}
\author{
\strong{Maintainer}: Pablo Fuenzalida \email{pablofuenzalida427@gmail.com}

}
\keyword{internal}
